Extract structs from implants session.go
